// generated-types/models/WebSocketResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserNotificationDto } from './UserNotificationDto';
import {
    UserNotificationDtoFromJSON,
    UserNotificationDtoFromJSONTyped,
    UserNotificationDtoToJSON,
    UserNotificationDtoToJSONTyped,
} from './UserNotificationDto';

/**
 * WebSocket response message
 * @export
 * @interface WebSocketResponse
 */
export interface WebSocketResponse {
    /**
     * The type of WebSocket response
     * @type {string}
     * @memberof WebSocketResponse
     */
    responseType?: WebSocketResponseResponseTypeEnum;
    /**
     *
     * @type {UserNotificationDto}
     * @memberof WebSocketResponse
     */
    payload?: UserNotificationDto;
}


/**
 * @export
 */
export const WebSocketResponseResponseTypeEnum = {
    Notification: 'NOTIFICATION'
} as const;
export type WebSocketResponseResponseTypeEnum = typeof WebSocketResponseResponseTypeEnum[keyof typeof WebSocketResponseResponseTypeEnum];


/**
 * Check if a given object implements the WebSocketResponse interface.
 */
export function instanceOfWebSocketResponse(value: object): value is WebSocketResponse {
    return true;
}

export function WebSocketResponseFromJSON(json: any): WebSocketResponse {
    return WebSocketResponseFromJSONTyped(json, false);
}

export function WebSocketResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): WebSocketResponse {
    if (json == null) {
        return json;
    }
    return {

        'responseType': json['responseType'] == null ? undefined : json['responseType'],
        'payload': json['payload'] == null ? undefined : UserNotificationDtoFromJSON(json['payload']),
    };
}

export function WebSocketResponseToJSON(json: any): WebSocketResponse {
    return WebSocketResponseToJSONTyped(json, false);
}

export function WebSocketResponseToJSONTyped(value?: WebSocketResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'responseType': value['responseType'],
        'payload': UserNotificationDtoToJSON(value['payload']),
    };
}