// generated-types/models/WebSocketRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { NotificationPayload } from './NotificationPayload';
import {
    NotificationPayloadFromJSON,
    NotificationPayloadFromJSONTyped,
    NotificationPayloadToJSON,
    NotificationPayloadToJSONTyped,
} from './NotificationPayload';

/**
 * WebSocket request message
 * @export
 * @interface WebSocketRequest
 */
export interface WebSocketRequest {
    /**
     * The type of WebSocket request
     * @type {string}
     * @memberof WebSocketRequest
     */
    requestType?: WebSocketRequestRequestTypeEnum;
    /**
     * JWT token for authentication
     * @type {string}
     * @memberof WebSocketRequest
     */
    token?: string;
    /**
     * 
     * @type {NotificationPayload}
     * @memberof WebSocketRequest
     */
    payload?: NotificationPayload;
}


/**
 * @export
 */
export const WebSocketRequestRequestTypeEnum = {
    Register: 'REGISTER',
    Update: 'UPDATE'
} as const;
export type WebSocketRequestRequestTypeEnum = typeof WebSocketRequestRequestTypeEnum[keyof typeof WebSocketRequestRequestTypeEnum];


/**
 * Check if a given object implements the WebSocketRequest interface.
 */
export function instanceOfWebSocketRequest(value: object): value is WebSocketRequest {
    return true;
}

export function WebSocketRequestFromJSON(json: any): WebSocketRequest {
    return WebSocketRequestFromJSONTyped(json, false);
}

export function WebSocketRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): WebSocketRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'requestType': json['requestType'] == null ? undefined : json['requestType'],
        'token': json['token'] == null ? undefined : json['token'],
        'payload': json['payload'] == null ? undefined : NotificationPayloadFromJSON(json['payload']),
    };
}

export function WebSocketRequestToJSON(json: any): WebSocketRequest {
    return WebSocketRequestToJSONTyped(json, false);
}

export function WebSocketRequestToJSONTyped(value?: WebSocketRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'requestType': value['requestType'],
        'token': value['token'],
        'payload': NotificationPayloadToJSON(value['payload']),
    };
}