// generated-types/models/NotificationPayload.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Payload for notification status updates via WebSocket
 * @export
 * @interface NotificationPayload
 */
export interface NotificationPayload {
    /**
     * The new status for the notification
     * @type {string}
     * @memberof NotificationPayload
     */
    notificationStatus?: NotificationPayloadNotificationStatusEnum;
    /**
     * ID of the user notification to update
     * @type {number}
     * @memberof NotificationPayload
     */
    userNotificationID?: number;
}


/**
 * @export
 */
export const NotificationPayloadNotificationStatusEnum = {
    Unread: 'UNREAD',
    Read: 'READ'
} as const;
export type NotificationPayloadNotificationStatusEnum = typeof NotificationPayloadNotificationStatusEnum[keyof typeof NotificationPayloadNotificationStatusEnum];


/**
 * Check if a given object implements the NotificationPayload interface.
 */
export function instanceOfNotificationPayload(value: object): value is NotificationPayload {
    return true;
}

export function NotificationPayloadFromJSON(json: any): NotificationPayload {
    return NotificationPayloadFromJSONTyped(json, false);
}

export function NotificationPayloadFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationPayload {
    if (json == null) {
        return json;
    }
    return {
        
        'notificationStatus': json['notificationStatus'] == null ? undefined : json['notificationStatus'],
        'userNotificationID': json['userNotificationID'] == null ? undefined : json['userNotificationID'],
    };
}

export function NotificationPayloadToJSON(json: any): NotificationPayload {
    return NotificationPayloadToJSONTyped(json, false);
}

export function NotificationPayloadToJSONTyped(value?: NotificationPayload | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'notificationStatus': value['notificationStatus'],
        'userNotificationID': value['userNotificationID'],
    };
}